Upload of final project
Submitted version of practical3 for concurrent systems. This assignment used promela to model the dining philosophers problem. There is some mistake in the was the ltl (linear temporal logic) statement is implemented. I think it some something to do with the fact that all the processes are sharing the same "states" thinking and eating, a possible fix is to make an array, indexed by the philosophers id. That way each phil has its own unique thinking and eating states