Working Load Balancing Implementation
This scheduler will look for runnable processes. It finds the process that has been run the least amount of times and executes it. 

If there are multiple processes that have the same amount of executes, then the one which is next in the round robin order gets executed. This is achieved by keeping track of the process that has been executed last(lastRun), then starting our search for the next process at an index just after the last executed process(lastRun+1). 

When we come across a least run process, we update the variables lowest and lowestIndex. Because of the way we search through the table of processes using modulo arithmetic, the round robin scheduling scheme is implemented naturally, so if any processes have equal execution counts then we already have the process that is next in the round robin order.

The method getIndex() just checks if out index has fallen off the end of the array, if so it wraps it back around to the start.